Optimized Chunk Serialization logic and compression ratio (20KiB to 6KiB avg)

+++ VoxelEngine/Voxel/Chunk.cs
@@ -24,7 +24,7 @@ namespace VoxelEngine.Voxel
         public Block* Data;
         public byte* MinY;
         public byte* MaxY;
-        public int BlockCount;
+        public ushort BlockCount;
 
         public BlockMetadataCollection BlockMetadata;
